requirements.txt
rich>=13.0.0
PyYAML>=6.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  
mysql-connector-python>=8.0.0  
pyodbc>=5.0.0  # For SQL Server
python-dotenv>=1.0.0
prompt-toolkit>=3.0.0
typing-extensions>=4.0.0